try ingredient icicle
